pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tinder_bot"
version = "0.1.0"
description = "Automated Hinge opener generator using GPT-4o Vision"
authors = [
    {name = "Hinge Bot Developer", email = "developer@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "pyautogui>=0.9.53",
    "opencv-python>=4.6.0.66",
    "pillow>=9.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer[all]>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.0.249",
    "mypy>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src/tinder_bot"

[project.scripts]
tinder-bot = "tinder_bot.cli:app" 